backend/pyproject.toml
[tool.poetry]
name = "django-ecommerce-backend"
version = "0.1.0"
description = ""
authors = ["Essa25 <essamemon234@gmail.com>"]
readme = "README.md"
packages = [{include = "django_ecommerce_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
django-allauth = "^65.4.1"
djangorestframework-simplejwt = "^5.4.0"
pillow = "^11.1.0"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
